# docs/swagger.yaml
definitions:
...
  routes.PrivateToPublicReq:
    properties:
      publicKey:
        example: a 66-character public key
        maxLength: 66
        minLength: 66
        type: string
      requestBody:
        example: look at description for example
        type: string
      signature:
        description: |-
          signature without 0x prefix is broken into
          R: sig[0:63]
          S: sig[64:127]
        example: a 128 character string created when you signed the request with your
          private key or account handle
        maxLength: 128
        minLength: 128
        type: string
    required:
    - publicKey
    - requestBody
    - signature
    type: object
...
  routes.StatusRes:
    properties:
      status:
        example: status of the request
        type: string
    type: object
...
info:
  contact:
    name: Opacity Staff
    url: https://telegram.me/opacitystorage
  description: Opacity backend for file storage.
  license:
    name: OPACITY LIMITED CODE REVIEW LICENSE LICENSE.md
  termsOfService: https://opacity.io/terms-of-service
  title: Storage Node
  version: "1.0"
paths:
...
  /api/v2/public-share/convert:
    post:
      consumes:
      - application/json
      description: |-
        convert private file to a public shared one
        requestBody should be a stringified version of:
        {
        "fileHandle": "a deterministically created file handle",
        "fileSize": 543534,
        }
      parameters:
      - description: an object to do the conversion of a private file to a public
          one
        in: body
        name: PrivateToPublicReq
        required: true
        schema:
          $ref: '#/definitions/routes.PrivateToPublicReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.StatusRes'
        "400":
          description: 'bad request, unable to parse request body: (with the error)'
          schema:
            type: string
        "403":
          description: signature did not match
          schema:
            type: string
        "404":
          description: the data does not exist
          schema:
            type: string
      summary: convert private file to a public shared one
...
swagger: "2.0"